+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Update vega library to [v5.12.0](https://github.com/vega/vega/releases/tag/v5.12
 Update vega-lite library and examples to [v.4.12.1](https://github.com/vega/vega-lite/releases)
 - [#93](https://github.com/RandomFractals/vscode-vega-viewer/issues/93)
 Update vega-embed library to v6.8.0
+- [#94](https://github.com/RandomFractals/vscode-vega-viewer/issues/94)
+Refine vega view filename/title display
 
 ## v8.9.0 (2020-04-29)